@@ -12,63 +12,67 @@
   },
   "engines": {
     "node": "6.9.2",
-    "npm": "3.10.9"
+    "npm": "10.16.0"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.1.0",
-    "@angular/common": "^5.1.0",
-    "@angular/compiler": "^5.1.0",
-    "@angular/compiler-cli": "^5.1.0",
-    "@angular/core": "^5.1.0",
-    "@angular/forms": "^5.1.0",
-    "@angular/http": "^5.1.0",
-    "@angular/platform-browser": "^5.1.0",
-    "@angular/platform-browser-dynamic": "^5.1.0",
-    "@angular/router": "^5.1.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.0",
-    "@fortawesome/free-brands-svg-icons": "^5.1.0",
-    "@fortawesome/free-solid-svg-icons": "^5.1.0",
-    "@ngx-share/core": "^5.3.1",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
+    "@babel/polyfill": "^7.4.4",
+    "@fortawesome/fontawesome-svg-core": "^1.2.18",
+    "@fortawesome/free-brands-svg-icons": "^5.8.2",
+    "@fortawesome/free-solid-svg-icons": "^5.8.2",
+    "@ngx-share/core": "^7.1.2",
     "@types/file-saver": "^2.0.1",
-    "angular-datatables": "^5.0.0",
+    "angular-datatables": "^7.0.0",
     "angular2-image-upload": "^1.0.0-rc.0",
-    "angular2-ladda": "^1.2.2",
-    "angular2-recaptcha": "0.6.0",
-    "core-js": "^2.5.1",
-    "datatables.net": "^1.10.16",
-    "datatables.net-dt": "^1.10.16",
-    "express": "^4.16.2",
+    "angular2-ladda": "^2.0.1",
+    "angular2-recaptcha": "1.1.0",
+    "core-js": "^2.6.9",
+    "datatables.net": "^1.10.19",
+    "datatables.net-dt": "^1.10.19",
+    "express": "^4.17.1",
     "file-saver": "^2.0.2",
     "font-awesome": "^4.7.0",
-    "jquery": "^3.2.1",
-    "moment": "^2.19.1",
-    "ngx-bootstrap": "^2.0.0-beta.10",
-    "rxjs": "^5.5.2",
-    "tinymce": "^4.7.1",
+    "jquery": "^3.4.1",
+    "moment": "^2.24.0",
+    "ngx-bootstrap": "^4.2.0",
+    "rxjs": "^6.5.2",
+    "rxjs-compat": "^6.5.2",
+    "tinymce": "^4.9.4",
     "ts-helpers": "^1.1.2",
-    "typescript": "^2.6.2",
-    "zone.js": "^0.8.18"
+    "tslib": "^1.9.0",
+    "typescript": "3.4.5",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.4",
-    "@types/datatables.net": "^1.10.7",
-    "@types/jasmine": "2.8.4",
-    "@types/jquery": "^3.2.15",
-    "@types/node": "~9.3.0",
-    "codelyzer": "~4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "^8.0.1",
+    "@types/datatables.net": "^1.10.17",
+    "@types/jasmine": "3.3.13",
+    "@types/jquery": "^3.3.29",
+    "@types/node": "~12.0.4",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "karma-remap-istanbul": "^0.6.0",
-    "protractor": "^5.3.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "webdriver-manager": "12.0.6"
+    "protractor": "^5.4.2",
+    "ts-node": "~8.2.0",
+    "tslint": "~5.17.0",
+    "webdriver-manager": "12.1.4"
   }
 }
