@@ -1,21 +1,21 @@
-import {NgModule} from "@angular/core";
-import {CommonModule} from "@angular/common";
-import {EditMessageFormComponent} from "../components/edit-message-form/edit-message-form.component";
-import {EditMessageComponent} from "../components/edit-message/edit-message.component";
-import {CreateMessageComponent} from "../components/create-message/create-message.component";
-import {TinymceComponent} from "../components/tinymce/tinymce.component";
-import {MessageFormComponent} from "../components/message-form/message-form.component";
-import {RouterModule, Routes} from "@angular/router";
-import {MessagesComponent} from "../components/messages/messages.component";
-import {AuthGuardAdminService} from "../services/auth-guard-admin.service";
-import {MessageItemComponent} from "../components/message-item/message-item.component";
-import {SharedModule} from "./shared.module";
-import {CommentFormComponent} from "../components/comment-form/comment-form.component";
-import {CommentComponent} from "../components/comment/comment.component";
-import {MessageComponent} from "../components/message/message.component";
-import {AlertModule} from "ngx-bootstrap";
-import {LaddaModule} from "angular2-ladda";
-import {ShareModule} from "@ngx-share/core";
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {EditMessageFormComponent} from '../components/edit-message-form/edit-message-form.component';
+import {EditMessageComponent} from '../components/edit-message/edit-message.component';
+import {CreateMessageComponent} from '../components/create-message/create-message.component';
+import {TinymceComponent} from '../components/tinymce/tinymce.component';
+import {MessageFormComponent} from '../components/message-form/message-form.component';
+import {RouterModule, Routes} from '@angular/router';
+import {MessagesComponent} from '../components/messages/messages.component';
+import {AuthGuardAdminService} from '../services/auth-guard-admin.service';
+import {MessageItemComponent} from '../components/message-item/message-item.component';
+import {SharedModule} from './shared.module';
+import {CommentFormComponent} from '../components/comment-form/comment-form.component';
+import {CommentComponent} from '../components/comment/comment.component';
+import {MessageComponent} from '../components/message/message.component';
+import {AlertModule} from 'ngx-bootstrap';
+import {LaddaModule} from 'angular2-ladda';
+import {ShareModule} from '@ngx-share/core';
 
 const routes: Routes = [
     {path: '', component: MessagesComponent},
@@ -32,7 +32,7 @@ const routes: Routes = [
         SharedModule,
         AlertModule.forRoot(),
         RouterModule.forChild(routes),
-        ShareModule.forRoot(),
+        ShareModule,
         LaddaModule.forRoot({
             style: "expand-right",
             spinnerSize: 20,
